:white_check_mark: Fix SupervisedRoutine test cases and move squeeze and metric test cases to test_regression_routine.py
